+++ web/app/routes/sitemap[.]xml.ts
@@ -1,7 +1,12 @@
 // @ts-ignore
 import { routes } from "virtual:remix/server-build";
 import { generateSitemap } from "@nasa-gcn/remix-seo";
+import type { SEOHandle } from "@nasa-gcn/remix-seo";
 import type { LoaderFunctionArgs } from "@remix-run/node";
+import {
+	fetchAllPublishedPages,
+	fetchAllUsersName,
+} from "./functions/queries.server";
 
 export async function loader({ request }: LoaderFunctionArgs) {
 	return generateSitemap(request, routes, {
@@ -11,3 +16,24 @@ export async function loader({ request }: LoaderFunctionArgs) {
 		},
 	});
 }
+
+export const handle: SEOHandle = {
+	getSitemapEntries: async (request: Request) => {
+		const [pages, users] = await Promise.all([
+			fetchAllPublishedPages(),
+    fetchAllUsersName()
+  ]);
+
+  const pageEntries = pages.map((page) => ({
+    route: `/${page.user.userName}/page/${page.slug}`,
+    lastmod: page.updatedAt.toISOString(),
+  }));
+
+  const userEntries = users.map((user) => ({
+    route: `/${user.userName}`,
+    lastmod: user.updatedAt.toISOString(),
+  }));
+
+		return [...pageEntries, ...userEntries];
+	},
+};
